Refactor project structure by moving mainthread.h to the threading directory and updating include paths
